backend/requirement.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
flask==3.0.3
flask-cors==5.0.0
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
flask-login==0.6.3
itsdangerous==2.2.0
jinja2==3.1.4
mako==1.3.6
markupsafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlalchemy==2.0.36
typing-extensions==4.12.2
werkzeug==3.1.1
